backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pymysql==1.1.1
python-multipart==0.0.12
pydantic==2.10.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
bcrypt==4.0.1
pandas==2.2.0
openpyxl==3.1.2